+++ db-support/postgres/workload-stats-setup.sh
@@ -9,17 +9,20 @@ STATS_DIR="../stats_data"
 PG_CONN="-U $USER"
 SKIP_EXTENSIONS="false"
 SKIP_VACUUM="false"
+CLEANUP="false"
 
 show_help() {
     RET=$1
     echo "Usage: $0 <options>"
     echo "Allowed options:"
-    echo -e "-d | --dir\t<directory> specifies the directory to store/load the stats data files, defaults to '../stats_data'"
-    echo -e "-f | --force\tdelete existing instance of the database if necessary"
-    echo -e "-t | --target\t<db name> name of the stats database, defaults to 'stats'"
-    echo -e "--pg-conn\t<connection> connection string to the PostgreSQL server (server, port, user) if required, e.g. '-h localhost -U admin'"
-    echo -e "--no-ext\tdoes not install any extensions"
-    echo -e "--no-vacuum\tdoes not run VACUUM ANALYZE after the import"
+    echo -e "-d | --dir <directory>\tspecifies the directory to store/load the Stats data files, defaults to '../stats_data'"
+    echo -e "-f | --force\t\tdelete existing instance of the database if necessary"
+    echo -e "-t | --target <db name>\tname of the Stats database, defaults to 'stats'"
+    echo -e "--pg-conn <connection>\tconnection string to the PostgreSQL server (server, port, user) if required, e.g. '-h localhost -U admin'"
+    echo -e "--no-fkeys\t\tdoes not load foreign key indexes to the database (includes both foreign key constraints as well as the actual indexes)"
+    echo -e "--no-ext\t\tdoes not install any extensions"
+    echo -e "--no-vacuum\t\tdoes not run VACUUM ANALYZE after the import"
+    echo -e "--cleanup\t\tremove the data files after import"
     exit $RET
 }
 
@@ -60,6 +63,10 @@ while [ $# -gt 0 ] ; do
             ;;
         --no-vacuum)
             SKIP_VACUUM="true"
+            shift
+            ;;
+        --cleanup)
+            CLEANUP="true"
             shift
             ;;
         --help)
@@ -124,6 +131,11 @@ if [ $SKIP_VACUUM == "false" ] ; then
     psql $PG_CONN $DB_NAME -c "VACUUM ANALYZE;"
 else
     echo ".. Skipping vacuuming"
+fi
+
+if [ $CLEANUP == "true" ] ; then
+    echo ".. Removing data files"
+    rm -rf $TARGET_DIR
 fi
 
 echo ".. Done"